+++ R/add_qa_score.R
@@ -1,23 +1,38 @@
 #' Assign quality assessment score to each article
 #'
-#' @param articles data.frame  loaded from \code{load_epidata_raw} function
-#' @param params data.frame  loaded from \code{load_epidata_raw} function
-#' @param models data.frame  loaded from \code{load_epidata_raw} function
+#' @param articles data.frame loaded from \code{load_epidata} function
+#' @param ignore_errors logical; if \code{TRUE}, the function will assign QA scores
+#' where possible (i.e. where all answers to quality assessment questions 
+#' are not NA) and set the QA score to NAfor articles where all answers are NA. 
+#' If \code{FALSE}, an error is thrown instead.
 #' @details 
-#' @return article data.frame with an updated column containing QS score out of 100.
+#' We have used a bespoke 7 question quality assessment (QA) questionnaire to
+#' assess the quality of articles. The questions can be retrieved using the
+#' \code{qa_questions} function. The function assigns a QA score to each article
+#' as the number of questions answered 'yes' divided by the total number of
+#' questions answered (an answer might be NA if the question is not relevant to
+#' the article under consideration). 
+#' Articles with all NA answers are excluded from the QA unless \code{ignore_errors}
+#' is set to \code{TRUE}.
+#' @return a named list consisting of two elements. The first element of the 
+#' list is the article data.frame with an updated column containing three new columns:
+#' \code{qs_denominator} (total number of questions answered), \code{qs_numerator}
+#' (number of questions answered 'yes') and \code{qs_score} (QA score). The second
+#' element of the list (named errors) is a data.frame containing articles 
+#' with all NA answers.
+#' @seealso \code{\link{qa_questions}}
 #' @export
-#' @importFrom package function
 #' @examples
-#' 
+#' lassa <- load_epidata("lassa")
+#' lassa_qa <- assign_qa_score(lassa$articles)
+#' head(lassa_qa$articles[, c("qa_denominator", "qa_numerator", "qa_score")])
 assign_qa_score <- function(articles, ignore_errors = FALSE) {
   
   assert_articles(articles)
 
-  ## Check if everything is NA; test that actually works
+  errors <- NULL
   
-  question_cols <- c(
-    "qa_m1", "qa_m2", "qa_a3", "qa_a4", "qa_d5", "qa_d6", "qa_d7"
-  )
+  question_cols <- qa_questions()$qnames
   if (! all(question_cols %in% colnames(articles))) {
     msg1 <- "Not all QA questions are present in the data"
     msg2 <- "Did you forget to use load_epidata to load data?"
@@ -31,22 +46,25 @@ assign_qa_score <- function(articles, ignore_errors = FALSE) {
   if (any(all_nas)) {
     message(
       paste(
-        sum(all_nas), "articles have all NAs for QA questions. 
-            These are excluded for QA assessment, please check that this is correct."
+        sum(all_nas), "articles have all NAs for QA questions; please check that
+         this is expected. QA score for these articles is set to NA."
       )
     )
     errors <- articles[all_nas, ]
+    if (! ignore_errors) {
+      stop("Please correct the errors before proceeding", call. = FALSE)
+    }
   }
   n_non_nas <- nquestions - check
-  articles$qs_denominator <- n_non_nas
-  articles$qs_numerator <- apply(
+  articles$qa_denominator <- n_non_nas
+  articles$qa_numerator <- apply(
     answers, 1, 
     function(x) sum(x %in% c('yes', 'Yes', 'YES', "1"), na.rm = TRUE) 
   )
   ## Articles with all NAs will have NA score
-  articles$qs_score <- NA
-  articles$qs_score[! all_nas] <- articles$qs_numerator[! all_nas] / 
-    articles$qs_denominator[! all_nas]   
+  articles$qa_score <- NA
+  articles$qa_score[! all_nas] <- articles$qa_numerator[! all_nas] / 
+    articles$qa_denominator[! all_nas]   
   
   list(articles = articles, errors = errors)
 }
@@ -73,12 +91,13 @@ qa_questions <- function() {
     "Q7 Data: Issues accounted for"
   )
   notes <- c(
-    "Enough equations / reference to methodological papers / code to reproduce 
+    "Enough equations/reference to methodological papers/code to reproduce 
      the analysis",
     "subjective criteria",
     "are assumptions and parameter values used as inputs (e.g. the generation 
      time distribution for a branching process model) clearly stated?",
     "objective criteria",
+    "Data is clearly described and reproducible",
     "Data issues are discussed and acknowledged",
     "Data issues are accounted for"
   )