+++ R/assertions.R
@@ -43,3 +43,34 @@ assert_string <- function(x, message = "%s must be character string", name = dep
   }
   invisible(TRUE)
 }
+
+
+assert_data_frame <- function(x, message = "%s must be a data frame", name = deparse(substitute(x))) {
+  if (!is.data.frame(x)) {
+    stop(sprintf(message, name), call. = FALSE)
+  }
+  invisible(TRUE)
+}
+
+assert_params <- function(x, message = "%s must have a column named 'parameter_type'.
+                                        Did you load the data using load_epidata?", name = deparse(substitute(x))) {
+  assert_data_frame(x)
+  if (!"parameter_type" %in% colnames(x)) {
+    stop(sprintf(message, name), call. = FALSE)
+  }
+  invisible(TRUE)
+}
+
+assert_articles <- function(x, message = "%s must have columns
+   'first_author_first_name', 'first_author_surname', 'year_publication', 
+   'covidence_id'. Did you load the data using load_epidata?", 
+   name = deparse(substitute(x))) {
+  assert_data_frame(x)
+  cols_needed <- c(
+    "first_author_first_name", "first_author_surname", 
+    "year_publication", "covidence_id")
+  if (!all(cols_needed %in% colnames(x))) {
+    stop(sprintf(message, name), call. = FALSE)
+  }
+  invisible(TRUE)
+}