+++ NEWS.md
@@ -1,3 +1,11 @@
+# epireview 1.1.3
+
+* FEATURE: Add a function to assign QA scores based on answers to our questionnaire
+
+# epireview 1.1.2
+
+* FEATURE: Add a function to create pretty article labels
+
 # epireview 1.1.1
 
-* Add filter_df_for_metamean function
+* FEATURE: Add filter_df_for_metamean function